slight change of heart: move state.LocalState into clistate instead of terraform
